+++ panda/src/vision/webcamVideoCursorOpenCV.h
@@ -37,7 +37,7 @@ private:
   bool get_frame_data(const unsigned char *&r,
                       const unsigned char *&g,
                       const unsigned char *&b,
-                      ssize_t &x_pitch, ssize_t &y_pitch);
+                      int &x_pitch, int &y_pitch);
 
   CvCapture *_capture;
